apiのreadonly属性指定（POST時のID不要） apiのenvelope追加（GETでList取得時のenvelope)

+++ src/apis/trip.py
@@ -1,4 +1,8 @@
-from flask_restx import Namespace, fields, Resource, reqparse, inputs
+import json
+
+from flask import request
+from flask_restx import Namespace, fields, Resource, abort
+from marshmallow import pprint
 
 from src.apis.location import location
 from src.apis.transportation import transportation
@@ -8,12 +12,13 @@ from src.database import db
 trip_namespace = Namespace('trips', description='旅行プランの中で行程を表す')
 trip = trip_namespace.model('TripModel', {
     'id': fields.Integer(
+        readonly=True,
         required=False,
         description='行程のID',
         example=0
     ),
     'plan_id': fields.Integer(
-        required=False,
+        required=True,
         description='旅行プランのID（FK）',
         example=1
     ),
@@ -27,12 +32,12 @@ trip = trip_namespace.model('TripModel', {
     'departure_date': fields.DateTime(
         required=False,
         description='出発時刻',
-        example='2020-01-01T00:00:00+09:00'
+        example='2020-01-01T00:00:00'
     ),
     'arrival_date': fields.DateTime(
         required=False,
         description='到着時刻',
-        example='2099-12-31T23:59:59+09:00'
+        example='2099-12-31T23:59:59'
     ),
     'transportation': fields.Nested(transportation)
 })
@@ -41,7 +46,7 @@ trip = trip_namespace.model('TripModel', {
 @trip_namespace.route('/')
 class TripList(Resource):
     # TripModelを利用して結果をパースしてリストで返す
-    @trip_namespace.marshal_list_with(trip)
+    @trip_namespace.marshal_list_with(trip, envelope='trips')
     def get(self):
         """
         Trip一覧取得
@@ -51,7 +56,7 @@ class TripList(Resource):
         return TripModel.query.all()
 
     @trip_namespace.marshal_with(trip)
-    @trip_namespace.expect(trip, validate=True)
+    @trip_namespace.expect(trip)
     def post(self):
         """
         Trip登録
@@ -63,56 +68,44 @@ class TripList(Resource):
         - departure_date
         - arrival_date
         - transportation_id（FK）
-        ## 注意事項
-        idは登録時に自動採番されるため不要
-        FK表記のあるパラメタはFK制約のため、関連テーブルにレコードが存在しないとエラーとなる
         """
-        parser = reqparse.RequestParser()
-        parser.add_argument('plan_id', type=int, help='FK')
-        parser.add_argument('departure_location_id', type=int, help='FK')
-        parser.add_argument('destination_location_id', type=int, help='FK')
-        parser.add_argument('order', type=int)
-        parser.add_argument('departure_date', type=inputs.datetime_from_iso8601)
-        parser.add_argument('arrival_date', type=inputs.datetime_from_iso8601)
-        parser.add_argument('transportation_id', type=int)
+        schema = TripSchema()
+        req = json.loads(request.data)
+        trip_insert = schema.load(req)
+        pprint(trip_insert)
 
-        args = parser.parse_args()
-        trip_insert = TripModel(
-            args.plan_id,
-            args.departure_location_id,
-            args.destination_location_id,
-            args.order,
-            args.departure_date,
-            args.arrival_date,
-            args.transportation_id)
+        if trip_insert.id is not None:
+            return abort(400, "Already exist ID. Failed to INSERT")
+
         db.session.add(trip_insert)
         db.session.commit()
-        res = TripSchema().dump(trip_insert)
-        return res, 201
+
+        return trip_insert, 201
 
 
-@trip_namespace.route('/<int:id>')
-@trip_namespace.doc(params={'id': 'id of trip'})
+@trip_namespace.route('/<int:trip_id>')
+@trip_namespace.doc(params={'trip_id': 'id of trip'})
 class TripController(Resource):
     # TripModelモデルを利用して結果をパースして単体で返す
     @trip_namespace.marshal_with(trip)
-    def get(self, id):
+    def get(self, trip_id):
         """
         Trip詳細取得
         - {int:id}で指定された行程の情報を返す
         - 行程情報には出発地、到着地、交通手段がネスト構造で返却される
         """
-        # ただし1個も見つからなかったら404を返す
-        return TripModel.query.filter(TripModel.id == id).first_or_404()
+        return TripModel.query.get(trip_id) or abort(404)
 
-    def delete(self, id):
+    def delete(self, trip_id):
         """
         Trip削除
         - {int:id}で指定された行程を削除する
         - FK制約の影響は未確認
         """
-        # 見つからなかったときの処理してないけど許して
-        target_trip = TripModel.query.filter(TripModel.id == id).first()
+        target_trip = TripModel.query.get(trip_id)
+        if target_trip is None:
+            return abort(404)
+
         db.session.delete(target_trip)
         db.session.commit()
-        return {'message': 'Success'}, 200
+        return {'message': 'Delete Success'}, 200
